+++ Models/ProgressRecord.cs
@@ -8,8 +8,8 @@ namespace RunClubAPI.Models
         public int ProgressRecordId { get; set; }
         public int UserId { get; set; }
         public DateTime ProgressDateTime { get; set; }
-        public int DistanceCovered { get; set; }
-        public int TimeTaken { get; set; }
+        public double DistanceCovered { get; set; }
+        public double TimeTaken { get; set; }
 
         // ✅ Allow User to be null by making it nullable
         public User? User { get; set; }
